feat(Spotify): Add Class to handle Spotify Objects (#1)
* feat(API): Impliment Real Spotify Class

Correctly Implements Spotify class and a search method.

* feat(Searching): Update Artist and Album Searching

Provides better handling of Artist and Album searching, defaulting to US

* feat(Spotify): Improve search functionality

Refactor the `Search` method in the `Spotify` class to accept additional parameters for search type and market. This allows for more specific and targeted searches, defaulting to the US market. The changes in this commit enhance the handling of Artist and Album searching.

* refactor(Spotify): Implement an Interface for Spotify Class

Creates an interface for the spotify class

* refactor(Controllers): Fully implement Spotify interface

Controller file is now correctly using the ISpotify interface to the

* chore: Remove commented code

Removes code from Program.cs that was commented out

* feat(Spotify): Update ISpotify interface and related classes

Update the ISpotify interface and its related classes to include a new method for retrieving a track by its ID. This change ensures that the ISpotify interface is up to date with the latest requirements and allows for more comprehensive data retrieval from the Spotify API.

* feat(Spotify): Add method for searching albums

This commit adds a new method, `SearchAlbums`, to the `ISpotify` interface and implements it in the `Spotify` class. The method allows for searching albums based on a search query and returns a list of `AlbumDto` objects. This enhancement improves the search functionality of the application and provides more comprehensive data retrieval from the Spotify API.

* refactor: Refactor Biz Logic to Spotify Class

Puts logic for finding objects and building DTOs into the Spotify class,

* docs(Spotify): Add summary to the SEarchTracks method

* chore(Document): Write better documentation in the Spotify Class

---------

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>