feat(User): Add functionality for authenticating a Spotify User (#2)
* feat(User): Add logic to authenticate user

Adds OAuth Authentication handling for the Spotify API

* feat(SpotifyUser): Intializes User Class

Officially Intializes the Spotify User class, with interface.  Login and

* feat(User): Create Method to get recent tracks played

Adds new method to get the most recent tracks played by the user

* feat(Spotify-Data-Collector): Add SpotifyClient property to ISpotifyUser interface

This commit adds a new property, `SpotifyClient`, to the `ISpotifyUser` interface in the `Spotify-Data-Collector` project. The `SpotifyClient` property allows for storing and accessing the Spotify client object within the `ISpotifyUser` interface. This enhancement improves the functionality and flexibility of the `ISpotifyUser` interface.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* feat: Add SpotifyClient property to ISpotifyUser interface

This commit adds a new property, `SpotifyClient`, to the `ISpotifyUser` interface. It allows for storing and accessing the Spotify client object within the interface, improving its functionality and flexibility.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* feat: Update SpotifyUser class to remove tokenExpiryTime field

Remove the tokenExpiryTime field from the SpotifyUser class to simplify the code and improve clarity. The token expiry time is now calculated and stored directly in the TokenExpireTime property. This change enhances the overall structure and maintainability of the class.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* refactor(SpotifyUser): Update loginURI to use environment variable

Update the loginURI in the SpotifyUser class to use the environment variable "SPOTIFY_REDIRECT_URI" instead of a hardcoded value. This change improves the flexibility and security of the code by allowing the loginURI to be easily configured through environment variables.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* feat(Login): Add Logic to handle no token

Adds logic flow to redirect user to the route they were on if the

---------

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>