feat(API): Restructure API Design (#3)
* feat: Update Spotify-Data-Collector.csproj dependencies

Update the Spotify-Data-Collector.csproj file to include the latest versions of the Carter, dotenv.net, SpotifyAPI.web, and Swashbuckle.AspNetCore packages. This ensures that the project is using the most up-to-date dependencies and improves compatibility and functionality.

* feat: Update SpotifyUser.cs to handle Spotify client authentication and retrieval

Refactor the GetSpotifyClientAsync method in the SpotifyUser.cs file to handle Spotify client authentication and retrieval. The code now properly sets the SpotifyAccessCode and retrieves the Spotify client asynchronously. This enhancement improves the functionality and flexibility of the SpotifyUser class.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* feat: Update SpotifyUser.cs to handle Spotify client authentication and retrieval

* feat: added regstistration routes via carter

* feat: Refactor registration endpoints to use Carter

Refactor the RegistrationEndpoints.cs file to use the Carter library for defining and handling registration routes. This change improves the code structure and readability by leveraging the features and conventions provided by Carter.

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>

* feat: Refactor registration endpoints to use Carter

* Refactor registration endpoints to use Carter

* Refactor SpotifyObjectsEndpoints.cs to use Carter for defining and handling Spotify object routes

* update 200 OK responses on Spotify routes, and fix the Namespace for Tracks DTO

* refactor: update spotify endpoint responses

* refactor(endpoints): Add Status Messages to rest of endpoints

---------

Co-authored-by: David Coxsey <dcoxsey@choctawnation.com>